+++ src/Values/Identifier.php
@@ -20,6 +20,7 @@ declare(strict_types=1);
 namespace LaravelJsonApi\Spec\Values;
 
 use Illuminate\Support\Str;
+use LaravelJsonApi\Contracts\Schema\Relation;
 use LaravelJsonApi\Core\Document\Error;
 use LaravelJsonApi\Core\Document\ErrorList;
 use LaravelJsonApi\Spec\Specification;
@@ -45,19 +46,32 @@ class Identifier extends Value
     private $value;
 
     /**
+     * @var Relation|null
+     */
+    private ?Relation $relation = null;
+
+    /**
      * Identifier constructor.
      *
      * @param Specification $spec
      * @param Translator $translator
      * @param string $path
      * @param mixed $value
+     * @param Relation|null $relation
+     *      the relation the identifier is in.
      */
-    public function __construct(Specification $spec, Translator $translator, string $path, $value)
-    {
+    public function __construct(
+        Specification $spec,
+        Translator $translator,
+        string $path,
+        $value,
+        Relation $relation = null
+    ) {
         $this->spec = $spec;
         $this->translator = $translator;
         $this->path = $path;
         $this->value = $value;
+        $this->relation = $relation;
     }
 
     /**
@@ -113,16 +127,16 @@ class Identifier extends Value
             ));
         }
 
-        if ($errors->isEmpty() && !$this->spec->exists($this->value->type, $this->value->id)) {
-            $errors->push($this->translator->resourceDoesNotExist(
-                $this->doesNotExist()
-            ));
+        if ($errors->isEmpty() && $error = $this->validateTypeAndId()) {
+            $errors->push($error);
         }
 
         return $errors;
     }
 
     /**
+     * Validate just the type member.
+     *
      * @return Error|null
      */
     private function validateType(): ?Error
@@ -159,6 +173,8 @@ class Identifier extends Value
     }
 
     /**
+     * Validate just the id member.
+     *
      * @return Error|null
      */
     private function validateId(): ?Error
@@ -190,6 +206,29 @@ class Identifier extends Value
     }
 
     /**
+     * Validations to run when both the type and id are separately valid.
+     *
+     * @return Error|null
+     */
+    private function validateTypeAndId(): ?Error
+    {
+        if ($this->relation && !in_array($this->value->type, $this->relation->allInverse(), true)) {
+            return $this->translator->resourceTypeNotSupportedByRelationship(
+                $this->relation,
+                $this->pathForTypeAndId(),
+            );
+        }
+
+        if (!$this->spec->exists($this->value->type, $this->value->id)) {
+            return $this->translator->resourceDoesNotExist(
+                $this->pathForTypeAndId(),
+            );
+        }
+
+        return null;
+    }
+
+    /**
      * Is the identifier in a list of identifiers?
      *
      * @return bool
@@ -212,7 +251,7 @@ class Identifier extends Value
      *
      * @return string
      */
-    private function doesNotExist(): string
+    private function pathForTypeAndId(): string
     {
         if ($this->inList()) {
             return $this->path;
@@ -224,4 +263,5 @@ class Identifier extends Value
 
         return $this->path;
     }
+
 }