+++ src/Values/ToOne.php
@@ -19,6 +19,7 @@ declare(strict_types=1);
 
 namespace LaravelJsonApi\Spec\Values;
 
+use LaravelJsonApi\Contracts\Schema\Relation;
 use LaravelJsonApi\Core\Document\ErrorList;
 use LaravelJsonApi\Spec\Factory;
 use LaravelJsonApi\Spec\Translator;
@@ -38,9 +39,9 @@ class ToOne extends Value
     private Factory $factory;
 
     /**
-     * @var string
+     * @var Relation
      */
-    private string $name;
+    private Relation $relation;
 
     /**
      * @var mixed
@@ -57,20 +58,20 @@ class ToOne extends Value
      *
      * @param Translator $translator
      * @param Factory $factory
-     * @param string $name
+     * @param Relation $relation
      * @param string $path
      * @param mixed $value
      */
     public function __construct(
         Translator $translator,
         Factory $factory,
-        string $name,
+        Relation $relation,
         string $path,
         $value
     ) {
         $this->translator = $translator;
         $this->factory = $factory;
-        $this->name = $name;
+        $this->relation = $relation;
         $this->path = rtrim($path, '/');
         $this->value = $value;
     }
@@ -87,7 +88,8 @@ class ToOne extends Value
         if ($this->valid()) {
             return $this->data = !is_null($this->value->data) ? $this->factory->createIdentifierValue(
                 "{$this->path}/data",
-                $this->value->data
+                $this->value->data,
+                $this->relation,
             ) : null;
         }
 
@@ -134,7 +136,7 @@ class ToOne extends Value
             return $errors->push($this->translator->fieldExpectsToOne(
                 $this->parent(),
                 $this->member() ?: 'data',
-                $this->name
+                $this->relation->name()
             ));
         }